+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 This is a maintenaince release:
 
- * Added SetSaslCredentials wrapper for librdkafka rd_kafka_sasl_set_credentials.
+ * Added SetSaslCredentials. This new method (on the Producer, Consumer, and
+   AdminClient) allows modifying the stored SASL PLAIN/SCRAM credentials that
+	 will be used for subsequent (new) connections to a broker.
 
 
 ## v1.9.2